UI Improvements
-Increased Font Size
-Anchored UI for better scaling on different resolutions
-added explicit support for 1960x1080
-Changed Font Color to White
-Changed Font Style
-Added Reset Button to Restart game
-Unsubscribed scripts from events inside onDestroy(); because that's important.
-Renamed FreakoutCan to Can
