--- Hangman2450/Assets/Scripts/FreakOutCan.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 78c04e3bcd6cf5a44b9f8aecb4d50d11
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
