+++ Hangman2450/Assets/Scripts/ResetButton.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class ResetButton : MonoBehaviour
+{
+    public void Reset() {
+        SceneManager.LoadScene(SceneManager.GetActiveScene().name);
+    }
+}
