+++ Hangman2450/Assets/Scripts/ResetButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 296e8ee03427de444aa67e8ff5d62f4b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
